requirements.txt
celery==5.2.3
Django==4.1.7
gunicorn==20.0.4
djangorestframework==3.14.0
PyJWT==2.6.0
djangorestframework-simplejwt==5.2.2
django-celery-results==2.2.0
django-cors-headers==3.14.0
django-password-validators==1.7.1
drf-yasg==1.21.5
factory_boy==3.2.0
Faker==3.0.1
mock==4.0.2
psycopg2-binary==2.9.5
pydantic<2.0.0,>=1.8.1
boto3==1.26.120
django-storages==1.13.2
django-countries==7.5.1
whitenoise==6.4.0
django-filter==23.2
